Security fixes and misc
 - Changed OperationCommands id to be int. This because the id is
   actually an unsigned short. But Java does not handle unsigned values,
   therefore we need to save it as an int.
 - Added some minor "Coding life improvements(?)" by allowing both with
   and without "get" and "is" lower or capital letters in our service
   properties file.
 - Modified the DataOutputHandler slightly. Might need to make some test
   for this specifically later.
 - Added new dependency on Apache commons lang3, a lot of nice stuff
   there.
 - Total overhaul of the SecurityUtil. Both encrypt() and
   calculateHMAC() works as intended now.
 - Did not like the non-functional way the construction of messages was
   before so I created a generic Handler interface and created a
   Pipeline. This allowed a more cleaner and easier to follow code.
 - Added handlers for every step in our "remote message creation",
   HeaderHandler, PayloadHandler, EncryptionHandler and finally
   HMACHandler. Names being quite self-descriptive.
 - Added custom exceptions for encryption failures and a generic invalid sms exception.
 - Added/Fixed RemoteMessageSMS to use the new pipeline setup.
 - Added the new RemoteMessagePipeline, it has a validate function that
   checks the size if it is a SMS message pipeline.
 Notes:
  - Our protocol will not allow a payload greater than 61 bytes long,
    header 18 bytes long and hmac 32 bytes long. This because we're
    using AES 128-bit which is 16-block key. So, with the padding the
    only two sizes that is closest to the 140 byte long SMS restriction
    is 128 and 144. 144 is too long and therefore we have to make sure
    it stays around 128 and thus 61 bytes long payload maximum. That is
    if header and hmac is kept strictly at their lengths.

Resolves #14

+++ Common/src/main/java/io/miso/core/OperationCommand.java
@@ -4,19 +4,19 @@ import java.util.Arrays;
 import java.util.Optional;
 
 public enum OperationCommand {
-    OC_PING((short) 0x0001);
+    OC_PING(0x0001);
 
-    private final short id;
+    private final int id;
 
-    OperationCommand(final short id) {
+    OperationCommand(final int id) {
         this.id = id;
     }
 
-    public static Optional<OperationCommand> getFromId(final short id) {
+    public static Optional<OperationCommand> getFromId(final int id) {
         return Arrays.stream(OperationCommand.values()).filter(oc -> oc.id == id).findFirst();
     }
 
-    public short getId() {
+    public int getId() {
         return id;
     }
 }
